Moved test, add assert tests
