+++ smt/surrogate_models/tests/test_surrogate_model_examples.py
@@ -207,7 +207,7 @@ class Test(unittest.TestCase):
         plt.ylabel("y")
         plt.legend(["Training data", "Prediction"])
         plt.show()
-    
+
     def test_krg_mixed(self):
         import numpy as np
         import matplotlib.pyplot as plt
@@ -216,16 +216,16 @@ class Test(unittest.TestCase):
 
         xt = np.array([0.0, 1.0, 2.0, 3.0, 4.0])
         yt = np.array([0.0, 1.0, 1.5, 0.5, 1.0])
-        
+
         ##    Vartype example
-    
+
         # vartype =["cont","int",("cate",3),("cate",2)]
-        #"cont" means x1 continuous
-        #"int" means x2 integer
-        #"(cate, 3)" means x3,x4 & x5 are 3 levels of the same categorical variable
-        #"(cate, 2)" means x6 & x7 are 2 levels of the same categorical variable
-        
-        vartype = ["int"]     
+        # "cont" means x1 continuous
+        # "int" means x2 integer
+        # "(cate, 3)" means x3,x4 & x5 are 3 levels of the same categorical variable
+        # "(cate, 2)" means x6 & x7 are 2 levels of the same categorical variable
+
+        vartype = ["int"]
         sm = KRG(theta0=[1e-2], vartype=vartype)
         sm.set_training_values(xt, yt)
         sm.train()
@@ -240,82 +240,7 @@ class Test(unittest.TestCase):
         plt.ylabel("y")
         plt.legend(["Training data", "Prediction"])
         plt.show()
-        
-    
 
-    def test_krg_mixed_2D(self):
-        import numpy as np
-        import matplotlib.pyplot as plt
-        from smt.problems import Sphere
-        from smt.sampling_methods import LHS
-        from smt.surrogate_models import KRG
-        vartype = ["cont","int"]     
-        sm = KRG(vartype=vartype, print_prediction=False)
-        fun = Sphere(ndim=2)
-        xlimits_relaxed= sm._relax_limits(fun.xlimits,dim=2)
-        sampling = LHS(xlimits=xlimits_relaxed, criterion="m")
-        xt = sampling(20)
-        xt = sm._project_values(xt)
-        yt = fun(xt)
-        sm.set_training_values(xt, yt)
-        sm.train()
-
-        X = np.arange(xlimits_relaxed[0, 0], xlimits_relaxed[0, 1], 0.25)
-        Y = np.arange(xlimits_relaxed[1, 0], xlimits_relaxed[1, 1], 0.25)
-        X, Y = np.meshgrid(X, Y)
-        Z = np.zeros((X.shape[0], X.shape[1]))
-
-        for i in range(X.shape[0]):
-            for j in range(X.shape[1]):
-                Z[i, j] = sm.predict_values(
-                    np.hstack((X[i, j], Y[i, j])).reshape((1, 2))
-                )
-        x_out= sm._assign_labels(xt,fun.xlimits)
-       
-    def test_krg_mixed_4D(self):
-        import numpy as np
-        import matplotlib.pyplot as plt
-        from smt.problems import Sphere
-        from smt.sampling_methods import LHS
-        from smt.surrogate_models import KRG
-        vartype = ["cont",("cate",2),"int"]     
-        xlimits=[[-10,10],["blue","red"],[-10,-10]]
-        sm = KRG(vartype=vartype, print_prediction=False)
-        fun = Sphere(ndim=4)
-        xlimits_relaxed= sm._relax_limits(xlimits,dim=4)
-        sampling = LHS(xlimits=xlimits_relaxed, criterion="m")
-        xt = sampling(20)
-        xt = sm._project_values(xt)
-        yt = fun(xt)
-        sm.set_training_values(xt, yt)
-        sm.train()
-        x_out= sm._assign_labels(xt,xlimits)
-
-
-
-      
-    def test_krg_mixed_5D(self):
-        import numpy as np
-        import matplotlib.pyplot as plt
-        from smt.problems import Sphere
-        from smt.sampling_methods import LHS
-        from smt.surrogate_models import KRG
-        vartype = ["cont",("cate",3),"int"]     
-        xlimits=[[-10,10],["blue","red","green"],[-10,-10]]
-        sm = KRG(vartype=vartype, print_prediction=False)
-        fun = Sphere(ndim=5)
-        xlimits_relaxed= sm._relax_limits(xlimits,dim=5)
-        sampling = LHS(xlimits=xlimits_relaxed, criterion="m")
-        xt = sampling(20)
-        xt = sm._project_values(xt)
-        yt = fun(xt)
-        sm.set_training_values(xt, yt)
-        sm.train()
-        x_out= sm._assign_labels(xt,xlimits)
-
-
-
-             
     def test_kpls(self):
         import numpy as np
         import matplotlib.pyplot as plt
@@ -340,82 +265,6 @@ class Test(unittest.TestCase):
         plt.legend(["Training data", "Prediction"])
         plt.show()
 
- 
-    def test_kpls_mixed(self):
-        import numpy as np
-        import matplotlib.pyplot as plt
-
-        from smt.surrogate_models import KPLS
-
-        xt = np.array([0.0, 1.0, 2.0, 3.0, 4.0])
-        yt = np.array([0.0, 1.0, 1.5, 0.5, 1.0])
-    
-        vartype = ["int"]     
-        sm = KPLS(theta0=[1e-2], vartype=vartype)
-        sm.set_training_values(xt, yt)
-        sm.train()
-
-        num = 100
-        x = np.linspace(0.0, 4.0, num)
-        y = sm.predict_values(x)
-
-        plt.plot(xt, yt, "o")
-        plt.plot(x, y)
-        plt.xlabel("x")
-        plt.ylabel("y")
-        plt.legend(["Training data", "Prediction"])
-        plt.show()
-      
-
-
-    def test_kpls_mixed_2D(self):
-        import numpy as np
-        import matplotlib.pyplot as plt
-        from smt.problems import Sphere
-        from smt.sampling_methods import LHS
-        from smt.surrogate_models import KPLS
-        vartype = ["cont","int"]     
-        sm = KPLS(vartype=vartype, print_prediction=False)
-        fun = Sphere(ndim=2)
-        xlimits_relaxed= sm._relax_limits(fun.xlimits,dim=2)
-        sampling = LHS(xlimits=xlimits_relaxed, criterion="m")
-        xt = sampling(20)
-        xt = sm._project_values(xt)
-        yt = fun(xt)
-        sm.set_training_values(xt, yt)
-        sm.train()
-
-        X = np.arange(xlimits_relaxed[0, 0], xlimits_relaxed[0, 1], 0.25)
-        Y = np.arange(xlimits_relaxed[1, 0], xlimits_relaxed[1, 1], 0.25)
-        X, Y = np.meshgrid(X, Y)
-        Z = np.zeros((X.shape[0], X.shape[1]))
-
-        for i in range(X.shape[0]):
-            for j in range(X.shape[1]):
-                Z[i, j] = sm.predict_values(
-                    np.hstack((X[i, j], Y[i, j])).reshape((1, 2))
-                )
-        x_out= sm._assign_labels(xt,fun.xlimits)
-       
-    def test_kpls_mixed_4D(self):
-        import numpy as np
-        import matplotlib.pyplot as plt
-        from smt.problems import Sphere
-        from smt.sampling_methods import LHS
-        from smt.surrogate_models import KPLS
-        vartype = ["cont",("cate",2),"int"]     
-        xlimits=[[-10,10],["blue","red"],[-10,-10]]
-        sm = KPLS(vartype=vartype, print_prediction=False)
-        fun = Sphere(ndim=4)
-        xlimits_relaxed= sm._relax_limits(xlimits,dim=4)
-        sampling = LHS(xlimits=xlimits_relaxed, criterion="m")
-        xt = sampling(20)
-        xt = sm._project_values(xt)
-        yt = fun(xt)
-        sm.set_training_values(xt, yt)
-        sm.train()
-        x_out= sm._assign_labels(xt,xlimits)
-
     def test_kplsk(self):
         import numpy as np
         import matplotlib.pyplot as plt
@@ -439,84 +288,6 @@ class Test(unittest.TestCase):
         plt.ylabel("y")
         plt.legend(["Training data", "Prediction"])
         plt.show()
-
-
-    def test_kplsk_mixed(self):
-        import numpy as np
-        import matplotlib.pyplot as plt
-
-        from smt.surrogate_models import KPLSK
-
-        xt = np.array([0.0, 1.0, 2.0, 3.0, 4.0])
-        yt = np.array([0.0, 1.0, 1.5, 0.5, 1.0])
-    
-        vartype = ["int"]     
-        sm = KPLSK(theta0=[1e-2], vartype=vartype)
-        sm.set_training_values(xt, yt)
-        sm.train()
-
-        num = 100
-        x = np.linspace(0.0, 4.0, num)
-        y = sm.predict_values(x)
-
-        plt.plot(xt, yt, "o")
-        plt.plot(x, y)
-        plt.xlabel("x")
-        plt.ylabel("y")
-        plt.legend(["Training data", "Prediction"])
-        plt.show()
-      
-
-    def test_kplsk_mixed_2D(self):
-        import numpy as np
-        import matplotlib.pyplot as plt
-        from smt.problems import Sphere
-        from smt.sampling_methods import LHS
-        from smt.surrogate_models import KPLSK
-        vartype = ["cont","int"]     
-        sm = KPLSK(vartype=vartype, print_prediction=False)
-        fun = Sphere(ndim=2)
-        xlimits_relaxed= sm._relax_limits(fun.xlimits,dim=2)
-        sampling = LHS(xlimits=xlimits_relaxed, criterion="m")
-        xt = sampling(20)
-        xt = sm._project_values(xt)
-        yt = fun(xt)
-        sm.set_training_values(xt, yt)
-        sm.train()
-
-        X = np.arange(xlimits_relaxed[0, 0], xlimits_relaxed[0, 1], 0.25)
-        Y = np.arange(xlimits_relaxed[1, 0], xlimits_relaxed[1, 1], 0.25)
-        X, Y = np.meshgrid(X, Y)
-        Z = np.zeros((X.shape[0], X.shape[1]))
-
-        for i in range(X.shape[0]):
-            for j in range(X.shape[1]):
-                Z[i, j] = sm.predict_values(
-                    np.hstack((X[i, j], Y[i, j])).reshape((1, 2))
-                )
-        x_out= sm._assign_labels(xt,fun.xlimits)
-       
-    def test_kplsk_mixed_4D(self):
-        import numpy as np
-        import matplotlib.pyplot as plt
-        from smt.problems import Sphere
-        from smt.sampling_methods import LHS
-        from smt.surrogate_models import KPLSK
-        vartype = ["cont",("cate",2),"int"]     
-        xlimits=[[-10,10],["blue","red"],[-10,-10]]
-        sm = KPLSK(vartype=vartype, print_prediction=False)
-        fun = Sphere(ndim=4)
-        xlimits_relaxed= sm._relax_limits(xlimits,dim=4)
-        sampling = LHS(xlimits=xlimits_relaxed, criterion="m")
-        xt = sampling(20)
-        xt = sm._project_values(xt)
-        yt = fun(xt)
-        sm.set_training_values(xt, yt)
-        sm.train()
-        x_out= sm._assign_labels(xt,xlimits)
-
-
-
 
     def test_gekpls(self):
         import numpy as np
